serverless/cloud-run/cr-gpu/ollama-rag/apps/indexer/requirements.txt
google-cloud-alloydb>=0.4.0,<1.0.0
langchain-google-vertexai>=2.0.9,<3.0.0
langchain>=0.3.11,<1.0.0
langchain-community>=0.3.11,<1.0.0
pgvector>=0.3.6,<1.0.0
google-cloud-alloydb-connector[pg8000]>=1.6.0,<2.0.0
cloud-sql-python-connector>=1.15.0,<2.0.0
langchain-google-alloydb-pg>=0.8.0,<1.0.0
SQLAlchemy>=2.0.36,<2.1.0
google-cloud-storage>=2.19.0,<3.0.0rc1
PyMuPDF>=1.25.1,<1.50.0



